refactor(auth): restructure feature into professional app layout
