+++ backend/Dockerfile
@@ -17,4 +17,4 @@ COPY . .
 EXPOSE 8000
 
 # 7. Command to run the application
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
