+++ backend/app/db/base.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String, DateTime
 from sqlalchemy.sql import func
-from database import Base
+from app.db.session import Base
 
 class User(Base):
     __tablename__ = "users"
